Move tree example into main TableEditor scenario (deleting TreeTableExample.cs)
